requirements.txt
apprise==1.9.4
bidict==0.23.1
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
dnspython==2.7.0
eventlet==0.40.3
flask==3.1.1
flask-jwt-extended==4.7.1
flask-socketio==5.5.1
greenlet==3.2.4
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
markdown==3.9
markupsafe==3.0.2
oauthlib==3.3.1
psycogreen==1.0.2
psycopg2-binary==2.9.10
pyaml==25.7.0
pyjwt==2.10.1
python-engineio==4.12.2
python-socketio==5.13.0
pyyaml==6.0.2
requests==2.32.5
requests-oauthlib==2.0.0
simple-websocket==1.1.0
simpleeval==1.0.3
urllib3==2.5.0
voluptuous==0.15.2
werkzeug==3.1.3
wsproto==1.2.0
